requirements.txt
joblib>=1.4.0
pandas>=2.0.0
spacy>=3.7.0
scikit-learn>=1.1.3
python-dateutil>=2.8.2
python-Levenshtein>=0.24.0
txtai>=7.1.0
flask>=2.2.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
